+++ openlane/flows/__init__.py
@@ -21,3 +21,4 @@ as a number of built-in flows.
 from .flow import FlowError, FlowException, FlowProgressBar, Flow
 from .sequential import SequentialFlow
 from . import builtins
+from .cli import cloup_flow_opts