+++ Changelog.md
@@ -1,3 +1,11 @@
+# 2.0.0-a48
+
+* Created `openlane.flows.cloup_flow_opts`, which assigns a number of `cloup`
+  commandline options to an external function for convenience.
+* Moved handling of `last_run` inside `Flow.start`
+* Moved handling of volare, setting log level and threadpool count to `cloup_flow_opts`
+
+
 # 2.0.0-a47
 
 * Update Magic, Netgen, and Yosys